feat:  add drag and drop for boards and updating data on server (sinc order)

+++ src/app/protected/home/home.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
 import { ApiService } from '../../public/api';
 import { MatDialog } from '@angular/material/dialog';
 import { CreateBoardDialogComponent } from '../components/create-board-dialog-component/create-board-dialog.component';
@@ -16,14 +16,24 @@ export class HomeComponent implements OnInit {
   selectedBoardId: string = '';
   newColumnTitle: string = '';
 
-  constructor(private apiService: ApiService, private dialog: MatDialog) {}
+  constructor(
+    private apiService: ApiService,
+    private dialog: MatDialog,
+    private cdRef: ChangeDetectorRef
+  ) {}
+
   ngOnInit(): void {
     this.getBoards();
+    this.apiService.getBoards();
+    console.log(this.boards);
+    console.log('loaded');
+    this.cdRef.detectChanges();
   }
 
   getBoards(): void {
     this.apiService.getBoards().subscribe((data) => {
       this.boards = data;
+      console.log('loaded in get berds');
     });
   }
 
@@ -49,6 +59,7 @@ export class HomeComponent implements OnInit {
       }
     });
   }
+
   deleteBoard(boardId: string, boardTitle: string): void {
     const dialogRef = this.dialog.open(ConfirmModalComponent, {
       data: {
@@ -69,8 +80,10 @@ export class HomeComponent implements OnInit {
   getBoardColumns(boardId: string) {
     this.apiService.getColoms(boardId).subscribe((data) => {
       this.columns = data;
-      console.log('Boards:', this.boards);
+      console.log('coloms from server:', this.columns);
       console.log(this.columns.length);
+      this.columns = this.columns.sort((a, b) => a.order - b.order);
+      console.log('sorted columns', this.columns);
     });
   }
 
@@ -130,38 +143,40 @@ export class HomeComponent implements OnInit {
 
     dialogRef.afterClosed().subscribe((result) => {
       if (result) {
+        const indexToDelete = this.columns.findIndex(
+          (col) => col._id === columnId
+        );
+        if (indexToDelete === -1) return;
         this.apiService
           .deleteColumn(this.selectedBoardId, columnId)
           .subscribe(() => {
             console.log('column deleted');
-            this.getBoardColumns(this.selectedBoardId);
+            this.columns.splice(indexToDelete, 1);
+            this.columns.forEach((col, index) => {
+              col.order = index + 1;
+              this.apiService.updateColumnData(
+                col.boardId,
+                col._id,
+                col.title,
+                col.order
+              );
+            });
           });
       }
     });
   }
 
-  //@ts-ignore
   onColumbDrop(event: CdkDragDrop<any>) {
-    const droppedColumnIndex = event.previousIndex;
-    const droppedColumn = this.columns[droppedColumnIndex];
-    console.log(this.columns);
-    console.log(event);
-    console.log(this.columns);
-    console.log(event.previousIndex);
-    console.log(event.currentIndex);
-    console.log(droppedColumn);
     moveItemInArray(this.columns, event.previousIndex, event.currentIndex);
-    // Update the order of the columns in the database
-    // ...
-    // this.apiService
-    //       .updateColumnData(
-    //         this.selectedBoardId,
-    //         column._id,
-    //         column.title,
-    //         column.order
-    //       )
-    //       .subscribe(() => {
-    //         this.getBoardColumns(this.selectedBoardId);
-    //       });
+
+    this.columns.forEach((column, index) => {
+      column.order = index + 1;
+      this.apiService.updateColumnData(
+        column.boardId,
+        column._id,
+        column.title,
+        column.order
+      );
+    });
   }
 }
